fix transform bug
