+++ frontend/src/actions/__tests__/sample_color.test.ts
@@ -48,8 +48,8 @@ describe('sampleColor action registration', () => {
         registerSampleColorAction();
         const action = actions.get('sampleColor')!;
         action.handler({ x: 123, y: 456 });
-        // `merged` mode + null activeLayerId → layer_id sentinel -1.
-        expect(engine.post).toHaveBeenCalledWith('pick_color', { x: 123, y: 456, layer_id: -1 });
+        // `merged` mode + null activeLayerId → id sentinel -1.
+        expect(engine.post).toHaveBeenCalledWith('pick_color', { x: 123, y: 456, id: -1 });
     });
 
     it('onMove queues a pick at the screenToCanvas-converted coordinates', () => {
@@ -58,7 +58,7 @@ describe('sampleColor action registration', () => {
         const fakeEvent = { clientX: 999, clientY: 888 } as unknown as PointerEvent;
         action.onMove!({ x: 0, y: 0 }, fakeEvent, 0, 0);
         // screenToCanvas stub returns { x: 50, y: 60 }.
-        expect(engine.post).toHaveBeenCalledWith('pick_color', { x: 50, y: 60, layer_id: -1 });
+        expect(engine.post).toHaveBeenCalledWith('pick_color', { x: 50, y: 60, id: -1 });
     });
 });
