+++ frontend/src/tools/magic_wand.svelte.ts
@@ -34,7 +34,7 @@ export const magicWandTool: Tool = {
 
         const mode = selectionMode(e);
         ctx.engine.post('select_magic_wand', {
-            layer_id: app.activeLayerId,
+            id: app.activeLayerId,
             seed_x: Math.round(cx),
             seed_y: Math.round(cy),
             tolerance: magicWandSession.tolerance,
